NEW [options] automatic management of option arguments
Before this, the callback (opt_option_callback_t) had the
responsibility to update the argv index depending on its usage
of the option argument.

* Now arg is given to callback only if present, and the engine
handles option arguments missing or unexpected.
* The callback must update the argv index ONLY if it uses a
non-given option argument (arg==NULL) or if it does not use
a given option argument (arg!=NULL). Otherwise, update of index
is automatic.
* The user will describe the option argument thanks to the 'arg'
member of the opt_options_desc_t structure: if NULL, no argument
is required, and an error is raised if given. If starting with '[',
it is optional, otherwise, it is mandatory and an error is raised
if not given.
* This commit requires program using vlib/options.h to be updated,
by removing argv index update.