new [rbuf,avltree] set errno to 0 when returned value is NULL
As NULL values can be inserted in rbuf and in avltree, the get
functions can return NULL on success. When the returned value
is NULL, errno is set to 0, and when an error occurs, NULL
is returned with errno != 0.
When user inserts NULL values, he must test errno as below:
'if (avltree_find(tree, data) == NULL && errno != 0) perror("find");'